move to util module
